Refactoring#3:
* unite AbstractInMemoryRepository with AbstractCrudInMemoryRepository;
* extract CrudRepository<T>
