+++ src/main/java/ua/com/kl/cmathtutor/repository/BugRepository.java
@@ -1,15 +1,7 @@
 package ua.com.kl.cmathtutor.repository;
 
-import java.util.List;
-import java.util.Optional;
-
 import ua.com.kl.cmathtutor.domain.entity.Bug;
 
-public interface BugRepository {
+public interface BugRepository extends CrudRepository<Bug> {
 
-    List<Bug> findAll();
-
-    Optional<Bug> findById(Integer id);
-
-    Bug save(Bug bug);
 }
