+++ src/main/java/ua/com/kl/cmathtutor/repository/DepartmentRepository.java
@@ -1,15 +1,7 @@
 package ua.com.kl.cmathtutor.repository;
 
-import java.util.List;
-import java.util.Optional;
-
 import ua.com.kl.cmathtutor.domain.entity.Department;
 
-public interface DepartmentRepository {
+public interface DepartmentRepository extends CrudRepository<Department> {
 
-    List<Department> findAll();
-
-    Optional<Department> findById(Integer id);
-
-    Department save(Department department);
 }
