+++ src/main/java/ua/com/kl/cmathtutor/repository/EmployeeRepository.java
@@ -1,15 +1,7 @@
 package ua.com.kl.cmathtutor.repository;
 
-import java.util.List;
-import java.util.Optional;
-
 import ua.com.kl.cmathtutor.domain.entity.Employee;
 
-public interface EmployeeRepository {
+public interface EmployeeRepository extends CrudRepository<Employee> {
 
-    List<Employee> findAll();
-
-    Optional<Employee> findById(Integer id);
-
-    Employee save(Employee employee);
 }
