+++ src/main/java/ua/com/kl/cmathtutor/repository/inmemory/AbstractCrudInMemoryRepository.java
@@ -5,19 +5,34 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
+import org.apache.commons.lang3.SerializationUtils;
+
 import ua.com.kl.cmathtutor.domain.entity.IdContainer;
+import ua.com.kl.cmathtutor.repository.CrudRepository;
 
 public abstract class AbstractCrudInMemoryRepository<T extends Serializable & IdContainer>
-	extends AbstractInMemoryRepository<T> {
+	implements CrudRepository<T> {
 
+    private AtomicInteger idCounter = new AtomicInteger(1);
     private Map<Integer, T> entitiesById = new ConcurrentHashMap<>();
 
+    protected Integer selectId() {
+	return Integer.valueOf(idCounter.getAndIncrement());
+    }
+
+    protected T deepCopy(T entity) {
+	return SerializationUtils.roundtrip(entity);
+    }
+
+    @Override
     public List<T> findAll() {
 	return entitiesById.values().stream().map(this::deepCopy).collect(Collectors.toList());
     }
 
+    @Override
     public Optional<T> findById(Integer id) {
 	if (id == null) {
 	    return Optional.empty();
@@ -25,6 +40,7 @@ public abstract class AbstractCrudInMemoryRepository<T extends Serializable & Id
 	return Optional.ofNullable(entitiesById.get(id));
     }
 
+    @Override
     public T save(T entity) {
 	if (entity.getId().equals(0) || !entitiesById.containsKey(entity.getId())) {
 	    entity.setId(selectId());
